feat(runtime): add array release verifier regression and docs
