+++ src/runtime/CMakeLists.txt
@@ -12,21 +12,26 @@ set(RT_PUBLIC_HEADERS
   ${CMAKE_CURRENT_SOURCE_DIR}/rt.hpp
   ${CMAKE_CURRENT_SOURCE_DIR}/rt_math.h
   ${CMAKE_CURRENT_SOURCE_DIR}/rt_random.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/rt_string.h
   ${CMAKE_CURRENT_SOURCE_DIR}/rt_array.h
   ${CMAKE_CURRENT_SOURCE_DIR}/rt_heap.h
 )
 
-add_library(rt STATIC ${RT_SOURCES})
-target_include_directories(rt
+add_library(viper_runtime STATIC ${RT_SOURCES})
+target_include_directories(viper_runtime
   PUBLIC
     $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/src/runtime>
-    $<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}/viper>
+    $<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}/viper_runtime>
 )
 
 if(NOT WIN32)
-  target_link_libraries(rt PUBLIC m)
+  target_link_libraries(viper_runtime PUBLIC m)
 endif()
 
+target_compile_definitions(viper_runtime
+  PRIVATE $<$<CONFIG:Debug>:VIPER_RC_DEBUG=1>
+)
+
 install(FILES ${RT_PUBLIC_HEADERS}
-  DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/viper
+  DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/viper_runtime
 )
