+++ tests/unit/CMakeLists.txt
@@ -4,7 +4,7 @@
 
 set(VIPER_UNIT_SUPPORT_LIBS support)
 set(VIPER_UNIT_SUPPORT_LIBS ${VIPER_UNIT_SUPPORT_LIBS} PARENT_SCOPE)
-set(VIPER_RUNTIME_TEST_LIBS rt)
+set(VIPER_RUNTIME_TEST_LIBS viper_runtime)
 set(VIPER_RUNTIME_TEST_LIBS ${VIPER_RUNTIME_TEST_LIBS} PARENT_SCOPE)
 
 function(viper_add_unit_support_tests)
